Move CopyOnWriteArrayRef to mlir/Support/ADTExtras.h
+ Some minor tweaks
